Making the algorithm base class an actual base class and the GDPoptSolver a very thin thing that just deals with the _impl magic in init and solve
