+++ pyomo/contrib/gdpopt/loa.py
@@ -57,7 +57,7 @@ class _GDP_LOA_Solver(_GDPoptAlgorithm):
         logger = config.logger
 
         # We'll need these to get dual info after solving subproblems
-        add_constraint_list(self.parent.original_util_block)
+        add_constraint_list(self.original_util_block)
 
         (master_util_block,
          subproblem_util_block) = _get_master_and_subproblem(self, config)
@@ -68,36 +68,36 @@ class _GDP_LOA_Solver(_GDPoptAlgorithm):
         original_obj = self._setup_augmented_Lagrangian_objective(
             master_util_block)
 
-        self.parent._log_header(logger)
+        self._log_header(logger)
 
         # main loop
-        while self.parent.iteration < config.iterlim:
-            self.parent.iteration += 1
+        while self.iteration < config.iterlim:
+            self.iteration += 1
 
             # solve linear master problem
-            with time_code(self.parent.timing, 'mip'):
+            with time_code(self.timing, 'mip'):
                 oa_obj = self._update_augmented_Lagrangian_objective(
                     master_util_block, original_obj, config.OA_penalty_factor)
                 mip_feasible = solve_MILP_master_problem(master_util_block,
                                                          config,
-                                                         self.parent.timing)
-                self.parent._update_bounds_after_master_problem_solve(
+                                                         self.timing)
+                self._update_bounds_after_master_problem_solve(
                     mip_feasible, oa_obj, logger)
 
             # Check termination conditions
-            if self.parent.any_termination_criterion_met(config):
+            if self.any_termination_criterion_met(config):
                 break
 
-            with time_code(self.parent.timing, 'nlp'):
+            with time_code(self.timing, 'nlp'):
                 _fix_master_soln_solve_subproblem_and_add_cuts(
                     master_util_block, subproblem_util_block, config, self)
 
             # Add integer cut
-            with time_code(self.parent.timing, "integer cut generation"):
+            with time_code(self.timing, "integer cut generation"):
                 add_no_good_cut(master_util_block, config)
 
             # Check termination conditions
-            if self.parent.any_termination_criterion_met(config):
+            if self.any_termination_criterion_met(config):
                 break
 
     def _setup_augmented_Lagrangian_objective(self, master_util_block):
